requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
python-dotenv
mysql-connector-python
scipy
streamlit
requests
